requirements.txt
Django>=5.2.0
django-import-export>=3.3.1
django-crispy-forms>=2.0
django-filter>=23.5 